Move trigonometry to right place
