+++ core/include/scipp/core/element/trigonometry.h
