+++ core/test/CMakeLists.txt
@@ -11,7 +11,7 @@ add_executable(
   element_event_operations_test.cpp
   element_unary_operations_test.cpp
   element_geometric_operations_test.cpp
-  element_trigonometry_operations_test.cpp
+  element_trigonometry_test.cpp
   value_and_variance_test.cpp
   view_index_test.cpp
 )
