+++ variable/trigonometry.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 
 #include "scipp/common/constants.h"
-#include "scipp/core/element/trigonometry_operations.h"
+#include "scipp/core/element/trigonometry.h"
 #include "scipp/variable/transform.h"
 #include "scipp/variable/trigonometry.h"
 
@@ -122,6 +122,17 @@ Variable atan(Variable &&var) {
 
 VariableView atan(const VariableConstView &var, const VariableView &out) {
   transform_in_place(out, var, element::atan_out_arg);
+  return out;
+}
+
+Variable atan2(const Variable &y, const Variable &x) {
+  return transform<std::tuple<double, float>>(y, x, element::atan2);
+}
+
+VariableView atan2(const VariableConstView &y, const VariableConstView &x,
+                   const VariableView &out) {
+  transform_in_place<std::tuple<double, float>>(out, y, x,
+                                                element::atan2_out_arg);
   return out;
 }
